♻️🎪 Extract loader
