+++ apps/show/src/routes/_exhibitor.exposants.$token._config.stand._index/section-stand-configuration.tsx
@@ -14,7 +14,7 @@ import { Price } from "#price/price.js";
 import { StandSizePrice } from "#stand-size/price.js";
 import { ShowExhibitorStatus } from "@prisma/client";
 import { Link, useLoaderData } from "@remix-run/react";
-import type { loader } from "./route";
+import type { loader } from "./loader.server.js";
 
 export function SectionStandConfiguration() {
   const { exhibitor } = useLoaderData<typeof loader>();
